+++ src/main.rs
@@ -1,5 +1,8 @@
 // #![warn(missing_docs)]
 
+use core::str;
+use std::process::Command;
+
 use clap::{crate_authors, crate_version, App, Arg};
 mod ff;
 use ff::{mpeg, probe};
@@ -36,6 +39,10 @@ fn main() {
         )
         .get_matches();
 
+    // Check if ffprobe && ffmpeg are installed.
+    check_for_ffprobe().expect("Dependency not found:");
+    check_for_ffmpeg().expect("Dependency not found");
+
     let path = matches.value_of("file").unwrap();
     let verse = matches.value_of("verse").unwrap();
     let output_path = matches.value_of("output_path").unwrap();
@@ -43,4 +50,24 @@ fn main() {
     let (start_time, end_time) = probe::Root::new(path).verse(verse);
 
     mpeg::cut(start_time, end_time, path, output_path);
+}
+
+/// Returns Ok(()) if the ffprobe is installed otherwise an error is returned.
+fn check_for_ffprobe() -> Result<(), &'static str> {
+    let ffprobe = Command::new("ffprobe").arg("-version").output().unwrap();
+    if ffprobe.status.success() == true {
+        Ok(())
+    } else {
+        Err("ffprobe seems not to be installed on the system. Install ffprobe before continuing.")
+    }
+}
+
+/// Returns Ok(()) if the ffprobe is installed otherwise an error is returned.
+fn check_for_ffmpeg() -> Result<(), &'static str> {
+    let ffmpeg = Command::new("ffmpeg").arg("-version").output().unwrap();
+    if ffmpeg.status.success() == true {
+        Ok(())
+    } else {
+        Err("ffmpeg seems not to be installed on the system. Install ffmpeg before continuing.")
+    }
 }